[FEATURE]: Add useCamera hook and Video component
