+++ src/store/store.js
@@ -14,16 +14,31 @@ const useStore = create(
     preStartSecondparticipantPose: [],
     firstParticipantPose: [],
     secondParticipantPose: [],
+    firstVideo: [],
+    secondVideo: [],
     isChildFirstParticipant: false,
     isChildSecondParticipant: false,
     fistParticipantPreparation: false,
     secondParticipantPreparation: false,
     showVideo: true,
     count: 0,
+    vidoeNumber: 0,
 
     addCount: (item) =>
       set((state) => ({
         count: state.count + 1,
+      })),
+    addVideoNumber: (item) =>
+      set((state) => ({
+        vidoeNumber: state.vidoeNumber + 1,
+      })),
+    addfirstVideo: (item) =>
+      set((state) => ({
+        firstVideo: [item],
+      })),
+    addSecondVideo: (item) =>
+      set((state) => ({
+        secondVideo: [item],
       })),
 
     addPerson: (item) =>
@@ -138,12 +153,15 @@ const useStore = create(
         secondParticipantPose: [],
         preStartFirstParticipantPose: [],
         preStartSecondparticipantPose: [],
+        firstVideo: [],
+        secondVideo: [],
         isChildFirstParticipant: false,
         isChildSecondParticipant: false,
         fistParticipantPreparation: false,
         secondParticipantPreparation: false,
         showVideo: true,
         count: 0,
+        vidoeNumber: 0,
       })),
   }))
 );
